+++ src/main/java/ua/opu/dl/pizzeria/dao/OrderDao.java
@@ -12,7 +12,7 @@ public interface OrderDao {
 
 	void deleteOrder(Order order);
 
-	Order loadById(Integer id);
+	Order loadById(long id);
 	Order loadByPhone(String phone);
 
 	List<Order> loadAllOrders();
