+++ src/main/java/ua/opu/dl/pizzeria/service/OrderService.java
@@ -12,7 +12,7 @@ public interface OrderService {
 
     void deleteOrder(Order order);
 
-    Order loadById(Integer id);
+    Order loadById(long id);
     Order loadByPhone(String phone);
     List<Order> loadAllOrder();
 }
